Fix inverted return value in `validateFile`
`validateFile()` currently returned whether there had been any errors. So the result was `true` if there were error messages and false when there were no error messages (or the validation was aborted by an exception).
The code in `main()` that was running the validation interpreted the results differently: `bool gotError = !validateFile(...);`

This **changes** the behavior of `validateFile()` such that it returns `true` on success and `false` on validation error (ie. `messageCount != 0`) or in case of an exception. This inverts the program's return code to be `0` for successful validation and `-1` in all other cases.